Create boilerplate for isomorphic redux apps
After wrestling with it for a day, I've come to terms with how this
system works.

The configurations for the server and client will be at the top level.
This can be expanded on in future. The actual code will be divided into
apps falling under 'src'. For example, there is an 'auth' app, where I
intend to write the basic 'logged in', 'logged out', 'register'
functionalities.

Another app will be created for things like 'blog' etc.

Of course, this is subject to change, but a start it is.